I-Ching
I-Ching is an entertainment app I wrote, which is now a featured
Facebook app. It features an application of javascript ajaxing, where
the user "tosses" three pennies six times, resulting in a broken or
unbroken line each time, until a hexagram is formed. A reading is
obtained after the sixth coin toss.